#7 refactoring: refactor apt
