+++ Server/carbonTracker/src/main/java/capstoneDesign/carbonTracker/apartment/dto/AptEnergyRequest.java
@@ -1,0 +1,9 @@
+package capstoneDesign.carbonTracker.apartment.dto;
+
+import lombok.Getter;
+
+@Getter
+public class AptEnergyRequest {
+    private String code;
+    private String date;
+}
