+++ Server/carbonTracker/src/main/java/capstoneDesign/carbonTracker/apartment/dto/AptListRequest.java
@@ -1,0 +1,10 @@
+package capstoneDesign.carbonTracker.apartment.dto;
+
+import lombok.Getter;
+
+@Getter
+public class AptListRequest {
+    private int code;
+    private int pageNum;
+    private int count;
+}
