--- Server/carbonTracker/src/main/java/capstoneDesign/carbonTracker/dto/AptEnergyRequest.java
@@ -1,6 +1,0 @@
-package capstoneDesign.carbonTracker.dto;
-
-public class AptEnergyRequest {
-    public String code;
-    public String date;
-}
