--- Server/carbonTracker/src/main/java/capstoneDesign/carbonTracker/dto/AptListRequest.java
@@ -1,7 +1,0 @@
-package capstoneDesign.carbonTracker.dto;
-
-public class AptListRequest {
-    public int code;
-    public int pageNum;
-    public int count;
-}
